+++ CruiserTerminal_NET2.1/CTerminal/CruiserTerminalPosition.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace CruiserTerminal.CTerminal
+{
+    internal class CruiserTerminalPosition : MonoBehaviour
+    {
+        //this class exist just to find gameobject
+    }
+}
