+++ CruiserTerminal_NET2.1/CTerminal/CruiserTerminalScript.cs
@@ -1,11 +1,9 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections;
 using Unity.Netcode;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
-namespace CruiserTerminal.Terminal
+namespace CruiserTerminal.CTerminal
 {
     public class CruiserTerminalScript : NetworkBehaviour, IHittable
     {
@@ -18,8 +16,15 @@ namespace CruiserTerminal.Terminal
         private bool punishment;
         private float penalty;
 
+        public bool cruiserTerminalInUse;
+        public InteractTrigger interactTrigger;
+        private Transform canvasMainContainer;
+
+        private PlayerActions playerActions;
+
         private Transform cruiserTerminal;
         private Transform cruiserTerminalPos;
+        private Transform terminal;
 
         bool IHittable.Hit(int force, Vector3 hitDirection, GameNetcodeStuff.PlayerControllerB playerWhoHit, bool playHitSFX, int hitID)
         {
@@ -32,7 +37,7 @@ namespace CruiserTerminal.Terminal
         }
 
         [ServerRpc]
-        void TerminalExplosionServerRPC(int force)
+        private void TerminalExplosionServerRPC(int force)
         {
             if (force <= 0 || !canBeHit || isDestroyed || !canDestroy)
                 return;
@@ -56,7 +61,7 @@ namespace CruiserTerminal.Terminal
         }
 
         [ClientRpc]
-        void TerminalExplosionClientRPC(bool punish)
+        private void TerminalExplosionClientRPC(bool punish)
         {
             StartCoroutine(TerminalMalfunction());
             if (punish)
@@ -83,6 +88,10 @@ namespace CruiserTerminal.Terminal
 
         private void Start()
         {
+            terminal = FindAnyObjectByType<Terminal>().transform.parent.parent;
+            cruiserTerminal = base.gameObject.transform;
+            cruiserTerminalPos = FindAnyObjectByType<CruiserTerminalPosition>().transform;
+
             maxHealth = CTConfig.maxHealth.Value;
             health = maxHealth;
             invTime = CTConfig.invTime.Value;
@@ -92,15 +101,54 @@ namespace CruiserTerminal.Terminal
             punishment = CTConfig.enablePenalty.Value;
             penalty = CTConfig.penalty.Value;
 
-            cruiserTerminal = base.gameObject.transform;
-            cruiserTerminalPos = FindAnyObjectByType<CruiserTerminalPosition>().transform;
-
+            cruiserTerminalInUse = false;
+            canvasMainContainer = terminal.Find("Canvas"); //not that bad as GameObject.Find() ig but this is Start() so not that much dif?
         }
 
         private void Update()
         {
             cruiserTerminal.position = cruiserTerminalPos.position;
             cruiserTerminal.rotation = cruiserTerminalPos.rotation;
+        }
+
+        public void BeginUsingCruiserTerminal()
+        {
+            playerActions.Movement.OpenMenu.performed += PressESC; //start listen esc key
+            cruiserTerminalInUse = true;
+        }
+
+        public void QuitCruiserTerminal()
+        {
+            playerActions.Movement.OpenMenu.performed -= PressESC; //stop listen esc key
+            cruiserTerminalInUse = false;
+            interactTrigger.StopSpecialAnimation();
+        }
+
+        private void PressESC(InputAction.CallbackContext context)
+        {
+            if (context.performed && cruiserTerminalInUse)
+            {
+                QuitCruiserTerminal();
+            }
+        }
+
+        private IEnumerator waitUntilFrameEndAndParent(bool active)
+        {
+            if (active)
+            {
+                canvasMainContainer.SetParent(cruiserTerminal);
+                canvasMainContainer.localPosition = new Vector3(-0.03f, 1.4f, 0.011f);
+                canvasMainContainer.localScale = new Vector3(0.004f, 0.0043f, 0.0016f);
+                canvasMainContainer.localRotation = Quaternion.Euler(new Vector3(0f, 90f, 0f));
+            }
+            else
+            {
+                canvasMainContainer.SetParent(terminal); //terminal
+                canvasMainContainer.localPosition = new Vector3(-0.516f, 0.284f, 1.284f);
+                canvasMainContainer.localScale = new Vector3(0.0015f, 0.0015f, 0.0016f);
+                canvasMainContainer.localRotation = Quaternion.Euler(new Vector3(0f, 78.0969f, 90f));
+            }
+            yield return new WaitForEndOfFrame();
         }
     }
 }
