--- CruiserTerminal_NET2.1/Terminal/CruiserTerminalPosition.cs
@@ -1,6 +1,0 @@
-﻿using UnityEngine;
-
-namespace CruiserTerminal.Terminal
-{
-    internal class CruiserTerminalPosition : MonoBehaviour { }
-}
